chore: basic ts skeleton for core
