+++ .eslintrc.js
@@ -1,57 +1,11 @@
 module.exports = {
-  parser: 'babel-eslint',
-  plugins: ['import'],
-  extends: 'airbnb',
-  env: {
-    commonjs: true,
-    es6: true,
-    jest: true,
-    node: true,
-    mocha: true,
-  },
-  parserOptions: {
-    ecmaVersion: 2018,
-  },
-  rules: {
-    'react/jsx-filename-extension': 'off',
-    'react/jsx-closing-bracket-location': 'off',
-    'jsx-a11y/no-static-element-interactions': 'off',
-    'jsx-a11y/click-events-have-key-events': 'off',
-    'react/forbid-prop-types': 'off',
-    'react/no-unescaped-entities': 'off',
-    'react/sort-comp': 'off',
-    'class-methods-use-this': 'off',
-    indent: ['error', 2],
-    'operator-linebreak': 'off',
-    'linebreak-style': ['error', 'unix'],
-    quotes: ['error', 'single'],
-    semi: ['error', 'always'],
-    'arrow-parens': 'off',
-    'no-plusplus': 'off',
-    'no-use-before-define': 'off',
-    'prefer-object-spread': 'off',
-    'object-curly-newline': 'off',
-    'implicit-arrow-linebreak': 'off',
-    'class-methods-use-this': 'off',
-    'import/no-unresolved': ['error', { caseSensitive: true, ignore: ['core/'] }],
-    'max-len': ['error', { code: 120, ignoreStrings: true, ignoreComments: true }],
-    'comma-dangle': [
-      'error',
-      {
-        arrays: 'always-multiline',
-        objects: 'always-multiline',
-        imports: 'always-multiline',
-        exports: 'always-multiline',
-        functions: 'ignore',
+  root: true,
+  extends: ['@arcblock/eslint-config-base'],
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
       },
-    ],
-    'no-param-reassign': [
-      'error',
-      {
-        props: false,
-      },
-    ],
-    'no-confusing-arrow': 'off',
-    'no-console': ['error', { allow: ['warn', 'error'] }],
+    },
   },
 };
