+++ package.json
@@ -11,17 +11,12 @@
     "url": "https://github.com/wangshijun"
   },
   "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint": "6.8.0",
-    "eslint-config-airbnb": "18.1.0",
-    "eslint-config-prettier": "latest",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jsx-a11y": "^6.1.1",
-    "eslint-plugin-prettier": "latest",
-    "eslint-plugin-react": "^7.11.0",
+    "eslint": "^8.21.0",
+    "@arcblock/eslint-config-base": "0.2.3",
+    "@arcblock/eslint-config-ts": "0.2.3",
     "husky": "^4.2.3",
     "jsdoc": "^3.6.10",
-    "lerna": "^3.22.1",
+    "lerna": "^5.5.0",
     "lint-staged": "^10.1.1",
     "prettier": "^2.0.2"
   },
