+++ packages/core/.eslintrc.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+module.exports = {
+  root: true,
+  extends: ['@arcblock/eslint-config-ts/base'],
+  parserOptions: {
+    project: path.resolve(__dirname, 'tsconfig.eslint.json'),
+  },
+};
