--- packages/core/index.js
@@ -1,3 +1,0 @@
-module.exports = require('./lib/datastore');
-module.exports.Model = require('./lib/model');
-module.exports.PromisedDatastore = require('./lib/promise');
