+++ packages/core/package.json
@@ -25,7 +25,6 @@
   },
   "files": [
     "lib",
-    "index.js",
     "bower.json"
   ],
   "dependencies": {
@@ -41,13 +40,22 @@
     "commander": "1.1.1",
     "exec-time": "0.0.2",
     "mocha": "8.x",
-    "underscore": "~1.13.3"
+    "underscore": "~1.13.3",
+    "@arcblock/eslint-config-ts": "0.2.3",
+    "@types/node": "^17.0.45",
+    "eslint": "^8.21.0",
+    "typescript": "^4.7.4"
   },
   "scripts": {
     "precommit": "npm run test",
     "prepush": "npm run test",
-    "test": "./node_modules/.bin/mocha --reporter spec --timeout 10000"
+    "clean": "rm -fr lib",
+    "prebuild": "npm run clean",
+    "build": "tsc",
+    "build:watch": "npm run build -- -w",
+    "test": "mocha --reporter spec --timeout 10000"
   },
-  "main": "index",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
   "license": "SEE LICENSE IN LICENSE"
 }
