+++ packages/core/src/cursor.js
