+++ packages/core/src/hashIndex.js
