+++ packages/core/src/index.ts
@@ -1,0 +1,4 @@
+import Datastore from './datastore';
+import PromisedDatastore from './promise';
+
+export { PromisedDatastore, Datastore };
