+++ packages/core/src/model.js
