+++ packages/core/tsconfig.eslint.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src", "tests", ".eslintrc.js"]
+}
