+++ packages/core/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src"],
+  "exclude": ["node_modules", "tests"],
+  "compilerOptions": {
+    "outDir": "src"
+  }
+}
